feat(editor): render Expr2::Float
